feat: misc production server setup
Add database sync to server start
Add loading prod sequelize/db params
Add key setting for production

+++ sequelize/sync.js
@@ -1,4 +1,8 @@
 // const Sequelize = require('sequelize');
+const myArgs = process.argv.slice(2);
+const force = myArgs[0] === 'force' ? true : false;
+const alter = myArgs[0] === 'alter' ? true : false;
+
 const s = require('./index');
 // Testing Connection and Syncing Tables with Models
 s.authenticate()
@@ -9,5 +13,16 @@ s.authenticate()
     console.log('Unable to connect to the database:', err);
   })
   .then(() => {
-    s.sync({ force: true });
+    if (alter) {
+      s.sync({ alter: true });
+    }
+    else if (force) {
+      s.sync({ force: true });
+    }
+    else {
+      console.log('No CLI args passed. Performing default sync');
+      s.sync();
+    }
   });
+
+
